+++ .idea/misc.xml
@@ -45,6 +45,8 @@
         <entry key="..\:/Users/User/AndroidStudioProjects/Spiice/app/src/main/res/layout/activity_step4.xml" value="0.30085470085470084" />
         <entry key="..\:/Users/User/AndroidStudioProjects/Spiice/app/src/main/res/layout/activity_step5.xml" value="0.30085470085470084" />
         <entry key="..\:/Users/User/AndroidStudioProjects/Spiice/app/src/main/res/layout/activity_step6.xml" value="0.30085470085470084" />
+        <entry key="..\:/Users/User/AndroidStudioProjects/Spiice/app/src/main/res/layout/custom_toolbar.xml" value="0.30085470085470084" />
+        <entry key="..\:/Users/User/AndroidStudioProjects/Spiice/app/src/main/res/layout/dialog_info_note.xml" value="0.3" />
         <entry key="..\:/Users/User/AndroidStudioProjects/Spiice/app/src/main/res/layout/dialog_item.xml" value="0.3" />
         <entry key="..\:/Users/User/AndroidStudioProjects/Spiice/app/src/main/res/layout/dialog_viewing_note.xml" value="0.30085470085470084" />
         <entry key="..\:/Users/User/AndroidStudioProjects/Spiice/app/src/main/res/layout/fragment_add_note.xml" value="0.3" />
