+++ app/src/main/java/com/example/projectone/ui/authorization/LoginFragment.kt
@@ -1,4 +1,4 @@
-package com.example.projectone.ui.fragments
+package com.example.projectone.ui.authorization
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -8,6 +8,7 @@ import android.widget.Toast
 import androidx.fragment.app.Fragment
 import com.example.projectone.R
 import com.example.projectone.databinding.FragmentLoginBinding
+import com.example.projectone.ui.addnote.AddNoteFragment
 import com.example.projectone.utils.addTextWatcher
 import com.example.projectone.utils.isValid
