+++ app/src/main/java/com/example/projectone/ui/listnotes/ListOfNotesFragment.kt
@@ -1,25 +1,27 @@
-package com.example.projectone.ui.fragments
+package com.example.projectone.ui.listnotes
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import androidx.fragment.app.viewModels
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.projectone.R
 import com.example.projectone.databinding.FragmentListOfNotesBinding
 import com.example.projectone.model.Note
-import com.example.projectone.repositories.NotesRepository
-import com.example.projectone.ui.notesadapter.NotesAdapter
-import com.example.projectone.utils.deleteNote
+import com.example.projectone.ui.addnote.AddNoteFragment
+import com.example.projectone.ui.authorization.LoginFragment
+import com.example.projectone.ui.listnotes.dialogs.NotePlanningInfoDialog
+import com.example.projectone.ui.listnotes.notesadapter.NotesAdapter
+import com.example.projectone.ui.listnotes.dialogs.ViewingNoteBottomSheetDialog
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
 
 class ListOfNotesFragment : Fragment() {
 
     private lateinit var binding: FragmentListOfNotesBinding
 
-    private val repository = NotesRepository()
-
-    private val notesAdapter = NotesAdapter(::onClickNote)
+    private val viewModel: ListNotesViewModel by viewModels()
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -52,17 +54,40 @@ class ListOfNotesFragment : Fragment() {
 
         binding.listOfNotes.apply {
             layoutManager = LinearLayoutManager(requireContext())
-            adapter = notesAdapter.apply {
-                setList(repository.getListNotes())
-            }
+            adapter = NotesAdapter(
+                onClickNote = { note -> onClickNote(note) },
+                onClickInfo = { note -> onClickInfo(note) }
+            )
         }
+
+        viewModel.listNotes.observe(viewLifecycleOwner) {
+            (binding.listOfNotes.adapter as NotesAdapter).setList(it)
+        }
+
+        viewModel.getListNotes()
     }
 
     private fun onClickNote(note: Note) {
         ViewingNoteBottomSheetDialog.newInstance(note, ::deleteNote).show(childFragmentManager, "")
     }
 
+    private fun onClickInfo(note: Note) {
+        NotePlanningInfoDialog.newInstance(note).show(childFragmentManager, "")
+    }
+
     private fun deleteNote(note: Note) {
-        this.deleteNote(note, repository, notesAdapter)
+        MaterialAlertDialogBuilder(requireContext())
+            .setTitle(note.title)
+            .setMessage(note.message)
+            .setNegativeButton(getString(R.string.negative_button_cancel)) { _, _ -> }
+            .setPositiveButton(getString(R.string.positive_button_delete)) { _, _ ->
+                viewModel.deleteNote(note)
+                viewModel.listNotes.observe(viewLifecycleOwner) {
+                    (binding.listOfNotes.adapter as NotesAdapter).setList(it)
+                }
+                viewModel.getListNotes()
+            }
+            .setCancelable(false)
+            .show()
     }
 }
