+++ app/src/main/res/drawable/ic_share.xml
@@ -1,7 +1,7 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
     android:width="@dimen/dp_30"
-    android:height="30dp"
-    android:tint="@color/black"
+    android:height="@dimen/dp_30"
+    android:tint="@color/text_input_layout"
     android:viewportWidth="24"
     android:viewportHeight="24">
     <path
