--- app/src/main/res/layout/activity_add_note.xml
@@ -1,97 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:layout_gravity="top"
-    android:orientation="vertical">
-
-    <include
-        android:id="@+id/toolbarCustom"
-        layout="@layout/custom_toolbar" />
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="@dimen/dp_16"
-        android:fontFamily="@font/red_hat_display_medium"
-        android:text="@string/add_note"
-        android:textColor="@color/text"
-        android:textSize="@dimen/sp_25" />
-
-    <com.google.android.material.textfield.TextInputLayout
-        android:id="@+id/titleNoteInputLayout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="@dimen/dp_16"
-        android:layout_marginTop="@dimen/dp_39"
-        android:layout_marginEnd="@dimen/dp_16"
-        android:hint="@string/title"
-        android:textColorHint="@color/text_input_layout"
-        app:boxBackgroundColor="@color/bg_text_input_layout"
-        app:boxCornerRadiusBottomEnd="@dimen/dp_8"
-        app:boxCornerRadiusBottomStart="@dimen/dp_8"
-        app:boxCornerRadiusTopEnd="@dimen/dp_8"
-        app:boxCornerRadiusTopStart="@dimen/dp_8"
-        app:boxStrokeWidth="@dimen/dp_0"
-        app:endIconDrawable="@drawable/ic_baseline_edit_calendar_24"
-        app:endIconMode="custom"
-        app:errorEnabled="true"
-        app:errorTextColor="@color/red">
-
-        <com.google.android.material.textfield.TextInputEditText
-            android:id="@+id/titleNoteInputEditText"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:fontFamily="@font/public_sans"
-            android:inputType="text"
-            android:textSize="@dimen/sp_16" />
-
-    </com.google.android.material.textfield.TextInputLayout>
-
-    <com.google.android.material.textfield.TextInputLayout
-        android:id="@+id/messageNoteInputLayout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="@dimen/dp_16"
-        android:layout_marginTop="@dimen/dp_39"
-        android:layout_marginEnd="@dimen/dp_16"
-        android:hint="@string/message"
-        android:textColorHint="@color/text_input_layout"
-        app:boxBackgroundColor="@color/bg_text_input_layout"
-        app:boxCornerRadiusBottomEnd="@dimen/dp_8"
-        app:boxCornerRadiusBottomStart="@dimen/dp_8"
-        app:boxCornerRadiusTopEnd="@dimen/dp_8"
-        app:boxCornerRadiusTopStart="@dimen/dp_8"
-        app:boxStrokeWidth="@dimen/dp_0"
-        app:errorEnabled="true"
-        app:errorTextColor="@color/red">
-
-        <com.google.android.material.textfield.TextInputEditText
-            android:id="@+id/messageNoteInputEditText"
-            android:layout_width="match_parent"
-            android:layout_height="160dp"
-            android:fontFamily="@font/public_sans"
-            android:gravity="top"
-            android:inputType="textMultiLine"
-            android:maxLines="10"
-            android:minLines="2"
-            android:textSize="@dimen/sp_16" />
-
-    </com.google.android.material.textfield.TextInputLayout>
-
-    <androidx.appcompat.widget.AppCompatButton
-        android:id="@+id/buttonAddNote"
-        android:layout_width="match_parent"
-        android:layout_height="@dimen/dp_54"
-        android:layout_marginStart="@dimen/dp_56"
-        android:layout_marginTop="@dimen/dp_186"
-        android:layout_marginEnd="@dimen/dp_56"
-        android:background="@drawable/bg_button"
-        android:fontFamily="@font/public_sans_medium"
-        android:text="@string/add"
-        android:textAllCaps="false"
-        android:textColor="@color/text_button"
-        android:textSize="@dimen/sp_18" />
-
-</LinearLayout>
