--- app/src/main/res/layout/activity_list_of_notes.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:layout_gravity="top"
-    android:orientation="vertical">
-
-    <include
-        android:id="@+id/toolbarCustom"
-        layout="@layout/custom_toolbar" />
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/listOfNotes"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-    </androidx.recyclerview.widget.RecyclerView>
-
-</LinearLayout>
