+++ app/src/main/res/values/dimens.xml
@@ -53,4 +53,5 @@
     <dimen name="dp_4">4dp</dimen>
     <dimen name="dp_2">2dp</dimen>
     <dimen name="dp_30">30dp</dimen>
+    <dimen name="sp_14">14sp</dimen>
 </resources>
